Add a Slider component.

+++ src/settings.zig
@@ -13,7 +13,7 @@ pub const version = "Cubyz α 0.12.0";
 pub const highestLOD: u5 = 5;
 
 
-pub var entityDistance: i32 = 2;
+pub var entityDistance: u16 = 2;
 
 
 pub var fov: f32 = 70;
@@ -22,7 +22,7 @@ pub var mouseSensitivity: f32 = 1;
 
 pub var fogCoefficient: f32 = 15;
 
-pub var renderDistance: i32 = 4;
+pub var renderDistance: u16 = 4;
 pub var LODFactor: f32 = 2.0;
 
 pub var bloom: bool = true;
